@@ -19,14 +19,14 @@
     "path-browserify": "^1.0.1",
     "process": "^0.11.10",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.0.0",
-    "react-dom": "^17.0.1",
-    "react-structural-metadata-editor": "https://github.com/avalonmediasystem/react-structural-metadata-editor#avalon-7.7.1",
-    "react_ujs": "^2.4.4",
+    "react": "^19.1.0",
+    "react-bootstrap": "^2.10.9",
+    "react-dom": "^19.1.0",
+    "react-structural-metadata-editor": "https://github.com/avalonmediasystem/react-structural-metadata-editor#v3.0.0-beta.0",
+    "react-on-rails": "^14.2.1",
     "sass": "^1.65.1",
     "sass-loader": "^13.3.2",
-    "shakapacker": "7.0.2",
+    "shakapacker": "8.0.1",
     "stream-browserify": "^3.0.0",
     "style-loader": "^3.3.3",
     "terser-webpack-plugin": "5",
