@@ -2,25 +2,21 @@
 // like app/views/layouts/application.html.erb. All it does is render <div>Hello React</div> at the bottom
 // of the page.
 
-import React from 'react'
-import ReactDOM from 'react-dom'
-import PropTypes from 'prop-types'
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import PropTypes from 'prop-types';
 
 const Hello = props => (
   <div>Hello {props.name}!</div>
-)
+);
 
 Hello.defaultProps = {
   name: 'David'
-}
+};
 
 Hello.propTypes = {
   name: PropTypes.string
-}
+};
 
-document.addEventListener('DOMContentLoaded', () => {
-  ReactDOM.render(
-    <Hello name="React" />,
-    document.body.appendChild(document.createElement('div')),
-  )
-})
+const root = createRoot(document.body.appendChild(document.createElement('div')));
+root.render(<Hello name="React" />);
