@@ -14,9 +14,24 @@
  * ---  END LICENSE_HEADER BLOCK  ---
 */
 
+// By default, this is used by react_on_rails for rendering the React components.
+
 import ReactOnRails from 'react-on-rails';
-import EmbeddedRamp from '../components/embeds/EmbeddedRamp';
+
+import CollectionList from '../components/CollectionList';
+import CollectionCarousel from '../components/CollectionCarousel';
+import CollectionDetails from '../components/CollectionDetails';
+import Search from '../components/Search';
+import MediaObjectRamp from '../components/MediaObjectRamp';
+import ReactButtonContainer from '../components/ReactButtonContainer';
+import PlaylistRamp from '../components/PlaylistRamp';
 
 ReactOnRails.register({
-  EmbeddedRamp,
+  CollectionList,
+  CollectionCarousel,
+  CollectionDetails,
+  Search,
+  MediaObjectRamp,
+  ReactButtonContainer,
+  PlaylistRamp,
 });
